check si le groupe user est supp A un user basic

+++ liste/liste.php
@@ -15,7 +15,12 @@
             <th>Question</th>
             <th>Réponse</th>
             <th>Date</th>
-            <th>Actions</th>
+            <?php
+                if($_SESSION['user']['id_usertype']>1){
+                    echo "<th>Actions</th>";
+                }
+            ?>
+            
         </tr>
         <?php
             foreach($faq as $question){
@@ -25,8 +30,10 @@
                 echo "<td>".$question["question"]."</td>";
                 echo "<td>".$question["reponse"]."</td>";
                 echo "<td>".$question["dat_question"]."</td>";
-                echo "<td><a href='editer.php'>Modifier</a></td>";
-                echo "<td><a href='supprimer.php'>Supprimer</a></td>";
+                if($_SESSION['user']['id_usertype']>1){
+                    echo "<td><a href='editer.php'>Modifier</a></td>";
+                    echo "<td><a href='supprimer.php'>Supprimer</a></td>";
+                }
                 echo "</tr>";
             }
         ?>
